FIXES #16: Change ITCH Message Protocol (4.1 -> 5.0)

+++ include/typedefs.h
@@ -15,7 +15,7 @@ constexpr size_t BUFFER_SIZE = 1024*1024+40;
 //arbitrary for now, our current data file is 8Gb so this will hold enough
 constexpr size_t MAX_BUFFERS = 10*1024;
 
-using ClientId = uint64_t;
+using ClientId = uint32_t;
 using Quantity = std::uint32_t;
 using Price = std::int64_t;
 using Time = uint64_t;
